pyproject.toml
[project]
name = "geminiliveagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "apscheduler>=3.10,<4.0",
    "beautifulsoup4>=4.14.3",
    "drissionpage>=4.1.1.2",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-genai>=1.0.0",
    "keyboard>=0.13.5",
    "lxml>=6.0.2",
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pygetwindow>=0.0.9",
    "python-dotenv>=1.0",
    "python-vlc>=3.0.21203",
    "pyyaml>=6.0.3",
]
